refactor: extract module walker to its own class
